% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionforlab5.r
\name{functionforlab5}
\alias{functionforlab5}
\title{Function for lab 5}
\usage{
functionforlab5(x)
}
\arguments{
\item{x}{any vector variable}
}
\value{
vector plus 95
}
\description{
adds 95 to a vector
}
\examples{
x=c(1,2,3,4), functionforlab5(x)
}
