% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mychisim2.R
\name{mychisim2}
\alias{mychisim2}
\title{Title}
\usage{
mychisim2(
  n1 = 10,
  n2 = 14,
  sigma1 = 3,
  sigma2 = 3,
  mean1 = 5,
  mean2 = 10,
  iter = 1000,
  ymax = 0.07,
  x = 40,
  y = 0.04,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
