% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{My Confidence interval
Finds 95% confidence interval for mean of sample d}
\usage{
myci(d, ...)
}
\arguments{
\item{d}{sample for which a confidence interval is being calculated}

\item{...}{}
}
\value{
confidence interval values
}
\description{
My Confidence interval
Finds 95% confidence interval for mean of sample d
}
\examples{
d=c(1,2,3,4,5), myci(d)
}
