% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myf.R
\name{myf}
\alias{myf}
\title{Square Function
Takes vector and returns a vector of squared components plus 1.}
\usage{
myf(x)
}
\arguments{
\item{x}{}
}
\value{
a vector of squared components plus one
}
\description{
Square Function
Takes vector and returns a vector of squared components plus 1.
}
\examples{
x=1:30;myf(x)
}
