% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFsim2.R
\name{myFsim2}
\alias{myFsim2}
\title{Title}
\usage{
myFsim2(
  n1 = 10,
  n2 = 14,
  sigma1 = 3,
  sigma2 = 2,
  mean1 = 5,
  mean2 = 10,
  iter = 1000,
  ymax = 0.9,
  x = 6,
  y = 0.5,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
