% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{Negative Binomial}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{# of trials until rth success}

\item{r}{# of desired successes}

\item{p}{probability of success}
}
\value{

}
\description{
Performs calculations for negative binomial distributions
}
