% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTsim.R
\name{myTsim}
\alias{myTsim}
\title{Title}
\usage{
myTsim(
  n1 = 10,
  sigma1 = 3,
  mean1 = 5,
  iter = 1000,
  ymax = 0.1,
  x = 2,
  y = 0.3,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
