% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTsim2.R
\name{myTsim2}
\alias{myTsim2}
\title{Title}
\usage{
myTsim2(
  n1 = 10,
  n2 = 14,
  sigma1 = 3,
  sigma2 = 3,
  mean1 = 5,
  mean2 = 10,
  iter = 1000,
  ymax = 0.5,
  x = 2,
  y = 0.4,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
